modifiy elapsed_ms type (from uint32_t to double) to handle more precise timing add types Clock, TimePoint and Duration examples have been slightly modified to reflect these changes

+++ examples/gravity.cpp
@@ -170,7 +170,7 @@ protected:
 		return true;
 	}
 
-	bool OnUserUpdate(uint32_t elapsed_ms) override {
+	bool OnUserUpdate(double elapsed_ms) override {
 		(void) elapsed_ms;
 		// handle user input
 		if (GetButton('q').pressed) return false;
